move _preprocess_text to TextProcessor
