--- tests/unit/process_text/extract_info/test_get_lowercase_text.py
@@ -1,8 +1,0 @@
-from src.process_text.extract_info import InformationExtractor
-
-
-def test_get_lowercase_text():
-    IE = InformationExtractor("dummy")
-    # pylint: disable=protected-access
-    lower = IE._get_lowercase_text("Fähigkeiten, test, gRosS ist blablablabla Waffen")
-    assert lower == "fähigkeiten, test, gross ist blablablabla waffen"
